+++ patch/autocore/status_c.js
@@ -1,4 +1,0 @@
-		    cpuinfo     = data[2],
-		    cpuusage    = data[3],
-		    tempinfo    = data[4],
-		    luciversion = data[5];