+++ patch/autocore/status_d.js
@@ -1,1 +1,0 @@
-			_('Architecture'),     cpuinfo.cpuinfo || boardinfo.system,